chore: remove third-party package
