+++ package.json
@@ -64,7 +64,6 @@
     "iso8601-duration": "^1.3.0",
     "lodash-es": "^4.17.21",
     "mixpanel-browser": "^2.45.0",
-    "ng-circle-progress": "^1.6.0",
     "ngx-color": "7.0.0",
     "rudder-sdk-js": "^2.3.0",
     "rxjs": "~6.6.7",
