+++ projects/observability/package.json
@@ -31,7 +31,6 @@
     "d3-shape": "^1.3.5",
     "d3-transition": "^1.2.0",
     "d3-zoom": "^1.8.3",
-    "ng-circle-progress": "^1.6.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
